CMV

Documentation

type CMVControl = Stream Bool

CMV control stream. When this is False, CMV does not count. When it goes to True, CMV exhales until the lower limit, then begins counting and runnings its cycle according to BPM and I:E ratio.

type SpontaneousBreath = Stream Bool

True when the patient has tried to inhale.

FIXME this is not CMV. Factor this better to be more clear.

data VMode

Organizational record for a ventilation mode: it says what the volume goal is and how much time is left to reach it.

Constructors

VMode

Fields

vmode_volume_ml :: Stream Word32

vmode_interval_us :: Stream Word32

cmv :: CMVControl -> SpontaneousBreath -> VMode

Continuous mandatory ventilation: fixed BPM and I:E ratio with either a pressure or volume goal.

TODO take streams as parameters rather than using "globals".

TODO pressure control support. TODO should respect PEEP, whether in VC or PC mode.

FIXME handles spontaneous breaths, so it's more general than CMV. Can specialize to CMV by giving constant False for the spontaneous breath stream.