Controls

Documentation

data Controls

Organizational record for all operator control signals (user input).

controls :: Controls

All controls are external (not derived).

type Mode = Word8

Macro ventilation mode (CMV, SIMV, etc.)

mode :: Stream Mode

mCMV :: Stream Mode

mSIMV :: Stream Mode

button_start_high_edge :: Stream Bool

True whenever the start button becomes pressed (changed, high leading edge).

button_stop_high_edge :: Stream Bool

Like button_start_high_edge but for the stop button.

error_controls :: Controls -> Stream Bool

True if there is an error in the control values. For instance, a ridiculously high or low volume limit, or a 0 in the I:E ratio.

TODO use an error code.

bpm_limited :: Stream Word8

Limit the breaths per minute to [6,40]

FIXME TODO determine whether these are sensible BPM limits.

ie_inhale_limited :: Stream Word8

The inhale portion of the I:E ratio is constrained by the exhale portion, and may not be 0.

ie_exhale_limited :: Stream Word8

The exhale part of the I:E ratio may be anything other than 0. The inhale portion will be constrained by this.

volume_limit_limited :: Stream Word32

Limits on the global volume limit. Unit is mL.

FIXME sensible limits. Currently it's 1L max and 100mL min

pressure_limit_limited :: Stream Word32

Limits on the global pressure limit. Unit is Pa.

FIXME sensible limits. Currently it's 5000Pa max and 100Pa min.

peep_limited :: Stream Word32

Positive end-expiratory pressure, in pascals.

Sane values for PEEP? It is relative to ambient pressure, and is positive by definition/name, so 0 is the obvious lower bound. The idea is that PEEP is always "on" but 0 means no PEEP.

TODO decide on a sensible upper bound. Apparently 10cmH2O is reasonable.

TODO use Int32 and guarantee it's positive and within bounds, so that other parts of the program do not need to unsafeCast.

cmv_mode :: Stream Bool

No sanity checking to do here, just give the control stream.

cmvVC :: Bool

cmvPC :: Bool

cmv_volume_goal_limited :: Stream Word32

In milliliters

One recommendation says that 6-8mL/kg ideal body weight is healthy. Since patient weights may vary, we'll allow for anywhere from 40-200kg. TODO check this

Does not use the global limited; those will be checked at all times and used to determine alarms.

cmv_pressure_goal_limited :: Stream Word32

Limited to 0-40 cmH2O

Does not use the global limited; those will be checked at all times and used to determine alarms.

global_max_flow :: Stream Word32

Max safe flow in uL/ms 90 L/minute TODO sensible value? This is also limited by hardware of course.

global_volume_min :: Stream Word32

mL

global_volume_max :: Stream Word32

mL

global_pressure_min :: Stream Word32

global_pressure_max :: Stream Word32