Kinematics

Description

Definitions related to the physical properties of the assembly which this software controls.

Documentation

forward_kinematics_mm :: Stream Double -> Stream Double

Given an angle, determine how far the piston has moved from zero, in millimeters.

The angle is in standard right-hand-rule style: 0 would mean the motor has pushed the piston all the way forward.

         / ---
       /  |    ---  L          ___________
  l  /    |y        ---       |
   /      |            ---    |
 /_ _ _ _ | _ _ _ _ _ _ _ _---|  bellows
    x_a         x_b           |
                              |___________

The input angle theta is the one on the left bottom.

reverse_kinematics :: Stream Double -> Stream Double

Given how far the piston lies to the right of the motor shaft, determine the angle in standard right-hand-rule (see forward_kinematics_mm which should invert this).

  \  ----
 |  \     ----   L
h|  l \         ----
 |      \  theta      ----      |
 |_ _ _ _ \ _ _ _ _ _ _ _ _ ----|
    x_a          x_b             |
              x

We're interested in the angle to the left of theta, call it phi. We'll compute that and subtract it from pi to get the radians, then convert to degrees.

The input is the length of the bottom edge to the right of the motor shaft, x_b. x_a is the length of the bottom of the triangle with h and l as other side lengths. x = x_a + x_b

We have

phi = acos(x_a/l)

(x_a + x_b)^2 + h^2 = L^2

h^2 = l^2 - x_a^2

From 2 and 3 we get

  x_a = (L^2 - l^2 - x_b^2) / 2*x_b

then we take the arccos and convert to degrees.

CAREFUL: this can be NaN if your start position does not make any sense given the other physical values l and big_l

l :: Stream Double

Length of the arm piece attached to the motor, in millimeters.

big_l :: Stream Double

Length of the arm piece attached to the bellows and to the other arm piece, in millimeters.

bellows_length_mm :: Double

How long is the bellows: how far can the piston move. theta ~= 30 should be the forward extremity.

bellows_diameter_mm :: Stream Double

The bellows is the thing that holds the air. This is its diameter in millimeters.

bellows_xsection_area_mm_2 :: Stream Double

Area of a circular cross section of the bellows. Will be used to compute volume. Pre-computed value is given.

start_position_mm :: Double

The distance in millimeters from the motor shaft to the piston when the piston is fully retracted

theta_max :: Stream Double

Max is when fully exhaled.

theta_min :: Stream Double

Min is when fully inhaled (cannot push any more air).

theta :: Stream Double

The current right-hand-rule angle, determined by way of the motor position encoder.

Will calculate often so this must be fast as possible.

We take the reverse kinematics at the start position, and then subtract the number of degrees the motor has turned, according to the encoder.

When the encoder is at 0, we want theta to be such that forward_kinematics_mm theta = 0 that's just to say, theta = reverse_kinematics 0

inverse_volume_delivered :: Stream Double -> Stream Double

Invert volume_delivered_mm_3: which theta would give that much volume.

TODO document and deal with edge cases. What if the input is more than could possibly be delivered?

volume_f :: Stream Double

flow_f :: Stream Int32 -> Stream Double

Estimate the rate of change in volume at an instant for a given motor velocity (L/s)

That's to say, the derivative w.r.t. time of

volume_delivered_mm_3 . theta

(theta is, though not explicitly written as such in this program text, a function of time, as all streams are)

For the sake of speed, we precompute the derivative of volume_delivered_mm_3 w.r.t. theta.

Alternatively, we could just compute the volume from theta every n milisecond and take the difference.

flow_f_observed :: Stream Word32 -> Stream Double -> Stream Double