Redundancy

Documentation

data WithRedundancy t

A principal stream and a redundant stream, with a stream of tolerances (probably constant, but allows for dynamic recalibration).

Constructors

WithRedundancy

Fields

principal :: Stream t

redundant :: Stream t

threshold :: Stream t

redundancy_check :: forall t. (Typed t, Num t, Ord t) => WithRedundancy t -> Stream Bool

A stream which is true whenever the principal and redundant streams differ by strictly more than the threshold.