Sensors

Documentation

sensors :: Sensors

Sensor streams, all of which are "external", i.e. must be provided by the hardware/firmware.

data Sensors

Organizational record for all sensor streams.

Constructors

Sensors

Fields

s_motor :: MotorSensors

s_pressure :: PressureSensors

s_oxygen :: OxygenSensors

s_flow :: FlowSensors

data MotorSensors

Organizational record for sensors relating to the piston and motor.

Constructors

MotorSensors

Fields

s_limit_low :: Stream Bool

The piston is at its lowest position (cannot pull any more air from the patient). Used for calibration (to 0 the encoder).

s_limit_high :: Stream Bool

The piston is at its highest position (cannot push any more air into the patient). TODO needed? Will we have this?

s_encoder_position :: Stream Int32

Signal for the encoder which tracks rotational position, from which the position of the piston can be inferred.

low_switch :: Stream Bool

The low switch for the motor must read True twice in a row in order to give True.

high_switch :: Stream Bool

Like low_switch.

encoder_position :: Stream Int32

Absolute encoder position. This is meaningless without reference. Use encoder_position_low and encoder_position_high, which are set when the low and high switches are triggered.

encoder_position_low :: Stream Int32

The encoder position at the low switch. Updated whenever the low switch is hit.

An initial calibration routine will be sure to hit the low switch and the high switch so as to set these streams to proper values.

encoder_position_high :: Stream Int32

data FlowSensors

Organizational record for sensors relating to flow, for patient-initiated (spontaneous) breaths.

Constructors

FlowSensors

Fields

s_insp_flow :: WithRedundancy Word32

s_exp_flow :: WithRedundancy Word32

s_air_in_flow :: WithRedundancy Word32

inhale_accumulator :: Stream Word32

Accumulated (averaged) inhale flow sensors.

exhale_accumulator :: Stream Word32

Accumulated (averaged) exhale flow sensors.

data PressureSensors

Organization record for sensors relating to pressure.

TODO units?

Constructors

PressureSensors

Fields

s_insp_pressure :: WithRedundancy Int32

insp_pressure_accumulator :: Stream Int32

How to filter the insp pressure sensor? Let's try smoothing it by taking the average of the next and the last signal.

inhale :: Stream Bool

True when an inhale has been detected: whenever inhale_accumulator is greater or equal to 8.

TODO good threshold

exhale :: Stream Bool

True when an exhale has been detected: whenever exhale_accumulator is greater or equal to 8.

accumulator :: (Num a, Eq a, Integral a, Typed a) => Stream a -> Stream a

data OxygenSensors

Organizational record for sensors relating to oxygen concentration.

Constructors

OxygenSensors

Fields

s_oxygen_concentration :: WithRedundancy Word32

flow_insp :: Stream Word32

flow_exp :: Stream Word32

pressure :: Stream Int32

oxygen :: Stream Word32

TODO infer from air in flow...

volume :: Stream Int32