Util

Documentation

clamp

Arguments

:: (Ord a, Typed a)

=> Stream a
Lower bound

-> Stream a
Upper bound

-> Stream a

-> Stream a

Ensure that a given stream is within the given bounds. Lower bound is checked first so if the lower bound exceeds the upper bound then the lower is given.

max :: (Ord a, Typed a) => Stream a -> Stream a -> Stream a

min :: (Ord a, Typed a) => Stream a -> Stream a -> Stream a

integral :: (Num a, Eq a, Typed a) => a -> Stream a -> Stream a

Integrate a stream with a given initial value.