Ventilator

Description

This is a copilot program designed to control a mechanical ventilatilation system.

The program works with data streams (type `Stream Int32` for instance), some of which are "extern" (set by a C program), and it defines "triggers", which correspond to C functions which will be called whenever a given condition (a `Stream Bool`) is true, and passed a set of values (sampled from the specified Streams).

A program specification is a set of triggers, their condition signals, and the arguments which they receive. This can be simulated in Haskell by giving mock values for the "extern" signals. It can also be compiled to C with C extern declarations for all "extern" signals as well as all declared triggers.

The specification for this program is called spec.

Documentation

pulsedata :: PulseData

Motor pulse data: converts the degrees-per-second velocity into the time between pulses, using hard-coded properties of the particular motor and driver.

motor_velocity :: Stream Int32

Calculate the desired motor velocity at an instant (degrees-per-second).

This is the essence of the ventilator program.

calibration_phase :: Stream Word8

calibration_phase_change :: Stream Bool

Is true whenever the current calibration phase differs from the previous one. Used to trigger function. Useful for debugging, to check the encoder value.

calibration_phase_next :: Stream Word8

alarm :: Stream Bool

Give True whenever human intervention is required.

As seen in spec, this signal controls when the C function raise_alarm will be called. That function would probably sound a buzzer and flash some lights.

TODO the alarm should give a code to indicate which signals are inconsistent, so that the UI signal can use it. Also should probably distinguish between critical alarms which imply motor shutdown, and warnings which do not.

TODO an alarm in case the high sensor is true but desired volume or pressure was not reached.

safe_to_move :: Stream Bool

The motor will not move unless this is true.

For now it's false whenever there is an alarm. Will need something better moving forward.

TODO FIXME is it better to have the kill switch mean bring the piston back to 0 and stay there until the operator intervenes? Would probably not be good to kill the motor with a high volume and pressure in the system.

spec :: Spec

The spec for the ventilator program.

gen_c :: IO ()

Write out the spec to C in ventilator.h and ventilator.c