Kinematics

Description

Definitions related to the physical properties of the assembly which this software controls.

Documentation

forward_kinematics_mm :: Stream Double -> Stream Double

Given an angle, determine how far the piston has moved from zero, in millimeters.

The angle is in standard right-hand-rule style: 0 would mean the motor has pushed the piston all the way forward.

         / ---
       /  |    -----  L        ___________
  l  /    |y_a        ------  |
   /      |                 H |
 /_ _ _ _ | _ _ _ _ _ _ _ _ _ |  bellows
    x_a         x_b           |
                              |___________

The input angle theta is the one on the left bottom.

l :: Stream Double

Length of the arm piece attached to the motor, in millimeters.

big_l :: Stream Double

Length of the arm piece attached to the bellows and to the other arm piece, in millimeters.

h :: Stream Double

How far above the rail is the joint between the bellows and the long arm (big_l), in millimeters

bellows_length_mm :: Double

How long is the bellows: how far can the piston move. theta ~= 30 should be the forward extremity.

bellows_diameter_mm :: Stream Double

The bellows is the thing that holds the air. This is its diameter in millimeters.

bellows_xsection_area_mm_2 :: Stream Double

Area of a circular cross section of the bellows. Will be used to compute volume. Pre-computed value is given.

start_position_mm :: Double

The distance in millimeters from the motor shaft to the piston when the piston is fully retracted

theta_max :: Stream Double

Max is when fully exhaled.

theta_min :: Stream Double

Min is when fully inhaled (cannot push any more air).

theta :: Stream Double

The current right-hand-rule angle, determined by way of the motor position encoder.

Will calculate often so this must be fast as possible.

We take the reverse kinematics at the start position, and then subtract the number of degrees the motor has turned, according to the encoder.

When the encoder is at 0, we want theta to be such that forward_kinematics_mm theta = 0 that's just to say, theta = reverse_kinematics 0

volume_f :: Stream Double

Volume in cubic millimiters.

volume_i :: Stream Int32

Volume as an integer in millilitres.