Motor

Description

Definitions related to the motor, its hardware driver, and rotary encoder.

Documentation

steps_per_rotation :: Stream Word32

How many steps per revolution. This will be known and static for the device, so is not really a configuration variable.

Found that 3200 does not work; the motor can't be made to move even at 90 degrees per second with this configuration. 1600 is fine though.

pulses_per_rotation :: Stream Word32

How many encoder pulses per revolution.

md_encoder_position :: Stream Int32 -> Stream Int32 -> Stream Int32

How many millidegrees between two encoder positions, with a sign according to the relative magnitudes (positive if the first is greater than the second).

TODO don't use millidegrees or radians, use instead subdivisions of Pi. The encoder divides 2Pi into 2048 equal arc lengths. If the encoder position is n greater than the low encoder position, then we have moved nPi/2048 radians. Ideally our trig computations can work in terms of Pi/2048 rather than degrees or radians.

type PulseData = Stream Int32

Microseconds per pulse, where the sign gives direction. 0 means no movement.

pulse_data_from_velocity_dps :: Stream Int32 -> PulseData

Given a desired degrees-per-second velocity, compute the microseconds per pulse and direction (its sign).

Microseconds per pulse is 0 if the speed is 0.

Uses the high and low switches to clamp the rate so that it does not move backwards when the low switch is on, for instance.

Enforces limits on speed according to the encoder, so that the piston assembly will not slam the endpoints and risk damage. Since the low and high switches set the encoder offsets, this also enforces 0 speed at the endpoints.

limit_motor_velocity :: Stream Int32 -> Stream Int32

If the machine is calibrated for the motor (high and low switches used to set encoder offsets high and low), then this function may be used to give a limit on the velocity, so as to ensure the motor does not slam the piston against the endpoints.